modify wing merge algorithm
Testing by Dark Visor found an edge case that the previous algorithm did not support (a Red 3 flight group in addition to an existing Red flight group with 2 fighters).  So, change the algorithm to merge any single-fighter flight group if there is an existing flight group where the number fits naturally into the wing sequence.